PHPCS fixes
